+++ changelog.md
@@ -1,5 +1,11 @@
 # Unreleased
 
+# 3.0.4
+
+- Add random command on koth win
+- Fix error with discord webhook on stop
+- Fix problem with placeholder on koth spawn
+
 # 3.0.3
 
 - Fix NMS